final project

+++ src/controllers/boardController.js
@@ -10,7 +10,7 @@ class boardController {
             // Đường dẫn file đã upload
             const cover = req.file.path;
             //goi den tang service
-            let dataBoard = { title, cover, };
+            let dataBoard = { title, cover };
             const board = await boardService.creat(dataBoard);
 
             res.status(200).json({"msg": "Tạo mới Board thành công !", board });
@@ -25,15 +25,22 @@ class boardController {
         try {
             const { title } = req.body;
             const { id } = req.params;
-
             let data = { title };
             const result = await boardService.update(id, data);
-            if (result) {
-                res.status(200).json({ 'msg': 'Updated' });
+            let checkBoardId = await boardService.checkBoardId(id);
+
+            if(!checkBoardId){
+                res.status(400).json({ 'msg': 'ID không tồn tại !' });
+            }else{
+                if (result) {
+                    res.status(200).json({ 'msg': 'Cập nhật thông tin Board thành công !' });
+                }
+                else {
+                    throw new Error('Cập nhật thông tin Board thất bại!')
+                }
             }
-            else {
-                throw new Error('Update fail !')
-            }
+
+            
 
         } catch (error) {
             throw error;
@@ -43,15 +50,22 @@ class boardController {
     //xóa board
     delete = async (req, res, next) => {
         try {
-            const { id } = req.params;
-            const result = await boardService.delete(id);
-            if (result) {
-                res.status(200).json({ 'msg': 'Deleted' });
-            }
-            else {
-                throw new Error('Delete fail !')
-            }
+            const  {id} = req.params;
+            let checkBoardId = await boardService.checkBoardId(id);
+           
+            if(!checkBoardId){
+                res.status(400).json({ 'msg': 'ID không tồn tại !' });
+            }else{
 
+                const result = await boardService.delete(id);
+                if (result) {
+                    res.status(200).json({ 'msg': 'Đã xóa thành công !' });
+                }
+                else {
+                    throw new Error('Xóa thất bại !')
+                }
+
+            }
 
         } catch (error) {
             throw error;
@@ -61,6 +75,7 @@ class boardController {
     //danh sách các board
     getAll = async (req, res, next) => {
         try {
+            
             const listBoards = await boardService.getAll();
             res.status(200).json({ listBoards });
 
@@ -68,6 +83,24 @@ class boardController {
             throw error;
         }
     }
+
+    //check boadID 
+ /*   checkBoardId = async (req, res, next) => {
+        try {
+            const boardId = req.params.id;
+
+            const result = await boardService.checkBoardId(boardId);
+            if(!result){
+                res.status(404).json({'msg': 'ID không tồn tại !'});
+            }
+            
+        } catch (error) {
+            throw error;
+        }
+    }
+
+     */
+
 
 }
